refactor: rename file `mail_server.py` -> `backend.py`
