+++ mail/mail/doctype/blocked_ip/blocked_ip.py
@@ -8,7 +8,7 @@ from frappe import _
 from frappe.model.document import Document
 from frappe.utils import now
 
-from mail.mail_server import get_mail_backend_api
+from mail.backend import get_mail_backend_api
 from mail.utils import extract_filter_values, rename_keys
 from mail.utils.dns import get_host_by_ip
