+++ mail/mail/doctype/mail_group/mail_group.py
@@ -5,7 +5,7 @@ import frappe
 from frappe import _
 from frappe.model.document import Document
 
-from mail.mail_server import MailBackendGroupManager
+from mail.backend import MailBackendGroupManager
 from mail.utils import normalize_email
 from mail.utils.cache import get_cluster_for_tenant, get_tenant_for_domain, get_tenant_for_user
 from mail.utils.user import has_role, is_system_manager, is_tenant_admin
